refactor: 리팩터링에 따른 테스트코드 수정, 테스트 코드 패키지 분리
