+++ src/test/java/kitchenpos/application/acceptance/menugroup/MenuGroupFactory.java
@@ -1,10 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.application.acceptance.menugroup;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.dto.MenuGroupRequest;
+import kitchenpos.dto.menugroup.MenuGroupRequest;
 import org.springframework.http.MediaType;
 
 public class MenuGroupFactory {
