+++ src/test/java/kitchenpos/application/acceptance/product/ProductFactory.java
@@ -1,29 +1,28 @@
-package kitchenpos.application;
+package kitchenpos.application.acceptance.product;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.domain.Menu;
-import kitchenpos.dto.MenuRequest;
+import kitchenpos.dto.product.ProductRequest;
 import org.springframework.http.MediaType;
 
-public class MenuFactory {
+public class ProductFactory {
 
-    public static ExtractableResponse<Response> 메뉴_생성_요청(MenuRequest params) {
+    public static ExtractableResponse<Response> 상품_생성_요청(ProductRequest params) {
         return RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(params)
-                .when().post("/api/menus")
+                .when().post("/api/products")
                 .then().log().all().
                 extract();
     }
 
-    public static ExtractableResponse<Response> 메뉴_조회_요청() {
+    public static ExtractableResponse<Response> 상품_조회_요청() {
         return RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/api/menus")
+                .when().get("/api/products")
                 .then().log().all().
                 extract();
     }
