+++ src/test/java/kitchenpos/application/unit/menu/MenuServiceTest.java
@@ -1,16 +1,19 @@
-package kitchenpos.application;
+package kitchenpos.application.unit.menu;
 
-import kitchenpos.domain.*;
-import kitchenpos.dto.MenuRequest;
-import kitchenpos.event.MenuCreatedEvent;
-import kitchenpos.repository.MenuGroupRepository;
-import kitchenpos.repository.MenuRepository;
+import kitchenpos.application.menu.MenuService;
+import kitchenpos.application.product.ProductService;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.product.Product;
+import kitchenpos.dto.menu.MenuRequest;
+import kitchenpos.repository.menu.MenuRepository;
+import kitchenpos.repository.menugroup.MenuGroupRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.context.ApplicationEventPublisher;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
@@ -30,7 +33,7 @@ class MenuServiceTest {
     MenuGroupRepository menuGroupRepository;
 
     @Mock
-    ApplicationEventPublisher eventPublisher;
+    ProductService productService;
 
     @DisplayName("메뉴를 등록한다.")
     @Test
@@ -45,7 +48,8 @@ class MenuServiceTest {
         후라이드치킨.addMenuProduct(후라이드치킨_상품);
 
         when(menuGroupRepository.findById(후라이드치킨.getMenuGroup().getId())).thenReturn(Optional.of(한마리메뉴));
-        MenuService menuService = new MenuService(menuRepository, menuGroupRepository, eventPublisher);
+        when(productService.getProduct(후라이드.getId())).thenReturn(후라이드);
+        MenuService menuService = new MenuService(menuRepository, menuGroupRepository, productService);
 
         Menu expectedMenu = mock(Menu.class);
         when(expectedMenu.getId()).thenReturn(1L);
@@ -57,7 +61,6 @@ class MenuServiceTest {
         // when
         Menu created_후라이드치킨 = menuService.create(MenuRequest.from(후라이드치킨));
         // then
-        verify(eventPublisher).publishEvent(any(MenuCreatedEvent.class));
         assertThat(created_후라이드치킨.getId()).isNotNull();
         assertThat(created_후라이드치킨.getName()).isEqualTo(후라이드치킨.getName());
         assertThat(created_후라이드치킨.getPrice()).isEqualTo(후라이드치킨.getPrice());
@@ -70,7 +73,7 @@ class MenuServiceTest {
         Menu menu = mock(Menu.class);
 
         when(menuRepository.findAll()).thenReturn(Arrays.asList(menu));
-        MenuService menuService = new MenuService(menuRepository, menuGroupRepository, eventPublisher);
+        MenuService menuService = new MenuService(menuRepository, menuGroupRepository, productService);
 
         // when
         List<Menu> menus = menuService.list();
