+++ src/test/java/kitchenpos/application/unit/order/OrderServiceTest.java
@@ -1,16 +1,14 @@
-package kitchenpos.application;
+package kitchenpos.application.unit.order;
 
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.dto.OrderLineItemRequest;
-import kitchenpos.dto.OrderRequest;
-import kitchenpos.event.OrderCreatedEvent;
-import kitchenpos.repository.MenuRepository;
-import kitchenpos.repository.OrderRepository;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.application.order.OrderService;
+import kitchenpos.application.order.OrderValidator;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.dto.order.OrderLineItemRequest;
+import kitchenpos.dto.order.OrderRequest;
+import kitchenpos.repository.order.OrderRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -31,9 +29,9 @@ class OrderServiceTest {
     @Mock
     OrderRepository orderRepository;
     @Mock
-    OrderTableRepository orderTableRepository;
-    @Mock
     ApplicationEventPublisher eventPublisher;
+    @Mock
+    OrderValidator orderValidator;
 
     @DisplayName("주문을 등록한다.")
     @Test
@@ -43,14 +41,13 @@ class OrderServiceTest {
 
         OrderTable orderTable = mock(OrderTable.class);
         when(orderTable.getId()).thenReturn(1L);
-        when(orderTableRepository.findById(orderTable.getId())).thenReturn(Optional.of(orderTable));
 
         OrderRequest orderRequest = new OrderRequest(orderTable.getId(), null, Arrays.asList(orderLineItemRequest));
 
         Order savedOrder = mock(Order.class);
         when(savedOrder.getId()).thenReturn(1L);
         when(orderRepository.save(any())).thenReturn(savedOrder);
-        OrderService orderService = new OrderService(orderRepository,  orderTableRepository, eventPublisher);
+        OrderService orderService = new OrderService(orderRepository, orderValidator, eventPublisher);
 
         // when
         Order createdOrder = orderService.create(orderRequest);
@@ -66,7 +63,7 @@ class OrderServiceTest {
         Order order = mock(Order.class);
         when(orderRepository.findAll()).thenReturn(Arrays.asList(order));
 
-        OrderService orderService = new OrderService(orderRepository,  orderTableRepository, eventPublisher);
+        OrderService orderService = new OrderService(orderRepository, orderValidator, eventPublisher);
 
         // when
         List<Order> orders = orderService.list();
@@ -88,7 +85,7 @@ class OrderServiceTest {
 
         when(orderRepository.findById(1L)).thenReturn(Optional.of(expectedOrder));
 
-        OrderService orderService = new OrderService(orderRepository,  orderTableRepository, eventPublisher);
+        OrderService orderService = new OrderService(orderRepository, orderValidator, eventPublisher);
         // when
         Order savedOrder = orderService.changeOrderStatus(1L, orderRequest);
         // then
